pyproject.toml
[project]
name = "opengamma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
]

[project.optional-dependencies]
torch = ["torch>=2.0.0", "sentence-transformers>=5.1.2"]
